extract binary-tree/codec pkg
