+++ binary-tree/README.md
@@ -26,4 +26,4 @@ Reference for `Morris traversal`
 - [Populating Next Right Pointers in Each Node](problems/next_right_1.go)
 - [Populating Next Right Pointers in Each Node II](problems/next_right_2.go)
 - [Lowest Common Ancestor of a Binary Tree](problems/lca.go)
-- [Serialize and Deserialize Binary Tree](problems/codec.go)
+- [Serialize and Deserialize Binary Tree](codec/codec.go)
