+++ binary-tree/codec/codec.go
@@ -1,4 +1,4 @@
-package problems
+package codec
 
 import (
 	"strconv"
